requirements.txt
# Python Dependencies for Driver Drowsiness Detection System

# Core ML/DL Libraries
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
numpy>=1.21.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.0.0

# Object Detection Frameworks
# detectron2 # Install from source: pip install 'git+https://github.com/facebookresearch/detectron2.git'

# Web Framework (for Backend API)
Flask>=2.0.0
Flask-CORS>=3.0.0
# FastAPI>=0.70.0  # Alternative to Flask
# uvicorn>=0.15.0  # For FastAPI

# Data Processing
pandas>=1.3.0
matplotlib>=3.3.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
PyYAML>=5.4.0
requests>=2.25.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Deployment
gunicorn>=20.1.0  # For production deployment

# Optional: GPU support
# Install CUDA-compatible PyTorch version based on your system
# For CUDA 11.6: torch==1.12.1+cu116 torchvision==0.13.1+cu116 -f https://download.pytorch.org/whl/torch_stable.html

# Environment Management
python-dotenv>=0.19.0

# Image Processing Extensions
scikit-image>=0.18.0
imageio>=2.9.0

# Model Deployment
onnx>=1.10.0  # For model conversion
onnxruntime>=1.9.0  # For ONNX inference

# Database (if needed)
# SQLAlchemy>=1.4.0
# sqlite3 (built-in)

# Logging and Monitoring
loguru>=0.5.0

# Performance Monitoring
psutil>=5.8.0
memory-profiler>=0.60.0